Finish Author Related Posts

+++ CMS/admin/includes/view_all_posts.php
@@ -24,13 +24,34 @@ if (isset($_POST['checkBoxArray'])){
                 $update_to_delete_status = mysqli_query( $connection, $query);
                 confirmQuery($update_to_delete_status);
                 break;
+
+            case 'clone':
+                $query = "SELECT * FROM posts WHERE post_id = '{$postValueId}' ";
+                $select_post_query = mysqli_query($connection, $query);
+
+                while ($row = mysqli_fetch_array($select_post_query)){
+
+                        $post_title = $row['post_title'];
+                        $post_author = $row['post_author'];
+                    $post_category_id = $row['post_category_id'];
+                    $post_status = $row['post_status'];
+                    $post_image = $row['post_image'];
+                    $post_tags = $row['post_tags'];
+                    $post_date = $row['post_date'];
+                    $post_content = $row['post_content'];
+                }
+                 $query = "INSERT INTO posts(post_category_id, post_title, post_author, post_date, post_image, post_content, post_tags, post_status) ";
+                 $query .= "VALUES({$post_category_id},'{$post_title}', '{$post_author}', now(), '{$post_image}', '{$post_content}' , '{$post_tags}', '{$post_status}') ";
+                 $copy_query = mysqli_query($connection, $query);
+                 if (!$copy_query){
+                     die("QUERY FAILED" . mysqli_error($connection));
+                 }
+                    break;
         }
 
     }
 }
 ?>
-
-
 
 <form action="" method="post">
 
@@ -42,6 +63,7 @@ if (isset($_POST['checkBoxArray'])){
             <option value="published">Publish</option>
             <option value="draft">Draft</option>
             <option value="delete">Delete</option>
+            <option value="clone">Clone</option>
         </select>
     </div>
 
@@ -65,12 +87,13 @@ if (isset($_POST['checkBoxArray'])){
         <th>View Post</th>
         <th>Edit</th>
         <th>Delete</th>
+        <th>Reset</th>
     </tr>
     </thead>
     <tbody>
 
     <?php
-    $query = "SELECT * FROM posts" ;
+    $query = "SELECT * FROM posts ORDER BY post_id DESC ";
     $select_post = mysqli_query($connection,$query);
     while($row = mysqli_fetch_assoc($select_post)){
         $post_id = $row['post_id'];
@@ -82,8 +105,7 @@ if (isset($_POST['checkBoxArray'])){
         $post_tags = $row['post_tags'];
         $post_comment_count = $row['post_comment_count'];
         $post_date = $row['post_date'];
-
-
+        $post_views_count = $row['post_views_count'];
     echo "<tr>";
     ?>
 
@@ -103,7 +125,6 @@ if (isset($_POST['checkBoxArray'])){
 }
 
 
-
     echo "<td>$post_status</td>";
     echo "<td><img width='200' height='90' src='../images/$post_image'></td>";
     echo "<td>$post_tags</td>";
@@ -111,7 +132,8 @@ if (isset($_POST['checkBoxArray'])){
     echo "<td>$post_date</td>";
     echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
     echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-    echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+    echo "<td><a onClick=\" javascript: return confirm('Are you shure you want to delete') \" href='posts.php?delete={$post_id}'>Delete</a></td>";
+    echo "<td><a href='posts.php?reset={$post_id}'>$post_views_count</a></td>";
     echo "</tr>"; 
 
     }
@@ -129,4 +151,12 @@ if (isset($_POST['checkBoxArray'])){
         $delete_query = mysqli_query($connection, $query);
         header("Location: posts.php");
     }
+
+
+if (isset($_GET['reset'])){
+    $the_post_id = $_GET['reset'];
+    $query = "UPDATE posts SET post_views_count = 0 WHERE post_id =" . mysqli_real_escape_string($connection, $_GET['reset'] . " " );
+    $resset_query = mysqli_query($connection, $query);
+    header("Location: posts.php");
+}
 ?>
